+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   - Don't require `T` to be `Copy`, only `Clone`.
   - Declare `const Q_MAX` as `u64` instead of `u32` as we use `u64` as the unsigned int type in other places.
   - Remove obsolete `'static` bound on `impl Linear for IntRange`.
+- Split normalization and quantization into different traits. This gives more flexibility.
 
 ## [0.2.0] - 2022-04-11
 ### Changed